requirements.txt
pandas>=2.2
numpy>=1.26
PyYAML>=6.0
matplotlib>=3.8
pyarrow>=15.0
pytest>=8.0
jupyter>=1.0
yfinance>=0.2
